requirements.txt
tensorflow==2.20.0
fastapi==0.109.0
uvicorn[standard]==0.27.0
pandas>=2.2.3
numpy>=2.1.0
scikit-learn>=1.6.0
pydantic>=2.5.3
joblib>=1.3.2
pyyaml>=6.0
pytest>=8.0.0
ruff>=0.3.0
httpx>=0.27.0
pytest-asyncio>=1.0.0
